Make omnibar scalable; add option to disable Total on omnibar
