+++ src/browser/graphics/components/przerwa/omnibar.tsx
@@ -2,6 +2,7 @@ import styled from 'styled-components';
 import { useState, useEffect } from 'react';
 import GenericMessage from '../omnibar/ticker/generic-message';
 import Milestones from '../omnibar/ticker/milestones';
+import Clock from '../clock';
 import { useReplicant } from 'use-nodecg';
 
 const OmnibarContainer = styled.div`
@@ -46,7 +47,7 @@ const Ticker = () => {
   function genericMsg(message: string) {
     return <GenericMessage message={message} onEnd={showNextElement} />;
   }
-  
+
   function showNextElement() {
     console.log('SHOWING NEXT MESSAGE');
     currentComponentIndex += 1;
@@ -63,7 +64,7 @@ const Ticker = () => {
 
   useEffect(() => {
     messageTypes.splice(0, messageTypes.length);
-    
+
     omnibarTextReplicant?.forEach(text => messageTypes.push(genericMsg(text)));
 
     if(omnibarEnableMilestones) {
@@ -87,36 +88,6 @@ const Ticker = () => {
       <span key={timestamp}>{currentElement}</span>
     </TickerContainer>
   );
-};
-
-const ClockDiv = styled.div`
-  font-size: 34px;
-  align-self: flex-end;
-  padding-left: 10px;
-  padding-right: 10px;
-  margin-top: auto;
-  margin-bottom: auto;
-  text-align: right;
-`;
-
-const Clock = () => {
-  const getClockHTML = () => {
-    const date_ob = new Date();
-    const hours = ('0' + date_ob.getHours()).slice(-2);
-    const minutes = ('0' + date_ob.getMinutes()).slice(-2);
-    return `${hours}:${minutes}`;
-  };
-
-  const [clock, setClock] = useState(getClockHTML());
-
-  useEffect(() => {
-    // update clock every half a second
-    setInterval(() => {
-      setClock(getClockHTML());
-    }, 500);
-  }, []);
-
-  return <ClockDiv dangerouslySetInnerHTML={{ __html: clock }} />;
 };
 
 export default BreakOmnibar;
